feat: refactor lazy loading module with separate feature files and global state management
